Simplify conditional privileges
Querying for an action by a user/role will now return either `True` or `False` or a given_condition.
Querying for an action by a user/role will no longer accept known_condition as an input, but rather return known_condition as an output - IF a known_condition is set for the combination

Remove orphaned packages from import statement.

Update README file accordingly.